backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
xrpl-py==3.0.0
python-jose[cryptography]==3.3.0
pydantic==2.6.1
httpx==0.24.1
python-dotenv==1.0.1
